Update windows installer

+++ background_installer.sh
@@ -142,15 +142,18 @@ WantedBy=multi-user.target" >> openplc.service
 if [ "$1" == "win" ]; then
     echo "Installing OpenPLC on Windows"
     cp ./utils/apt-cyg/apt-cyg ./
-    cp ./utils/apt-cyg/wget.exe /bin
+    if [ -f "/usr/bin/wget" ]
+    then
+        echo "found wget. Skipping install"
+    else
+        echo "wget not found. Installing from binary"
+        cp ./utils/apt-cyg/wget.exe /bin
+    fi
     install apt-cyg /bin
     apt-cyg update
-    # replace cygwin installed database to avoid installation errors
-    cp ./installed.db /etc/setup/installed.db
     apt-cyg install lynx
-    rm -f /bin/wget.exe
     # apt-cyg remove gcc-core gcc-g++ pkg-config automake autoconf libtool make python2 python2-pip sqlite3
-    apt-cyg install wget gcc-core gcc-g++ git pkg-config automake autoconf libtool make python2 python2-pip sqlite3 python3
+    apt-cyg install gcc-core gcc-g++ git pkg-config automake autoconf libtool make python2 python2-pip sqlite3 python3
     lynx -source https://bootstrap.pypa.io/pip/2.7/get-pip.py > get-pip.py
     lynx -source https://bootstrap.pypa.io/pip/get-pip.py > get-pip3.py
     /usr/bin/python2 get-pip.py
